+++ CHANGELOG.md
@@ -5,6 +5,8 @@ FlatLaf Change Log
 
 #### New features and improvements
 
+- Button and ToggleButton: Added missing border colors for pressed and selected
+  states. (issue #848)
 - Label: Support painting background with rounded corners. (issue #842)
 - Popup: Fixed flicker of popups (e.g. tooltips) while they are moving (e.g.
   following mouse pointer). (issues #832 and #672)